requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
# Tensorflow doesn't support CUDA 10 yet, (25 Jan) so nightly is needed
# tensorflow>=1.3.0
tf-nightly
tf-nightly-gpu
keras>=2.0.8
opencv-python
h5py
imgaug
IPython[all]